refactor(platform-ui): group platform implementation families
