+++ src/frontend/components/messagePresentation/components/messageListPlatform/messageListPlatform.android.ts
