+++ src/frontend/components/messagePresentation/components/messageListPlatform/messageListPlatform.ios.ts
