+++ src/frontend/components/messagePresentation/utils/builtInToolPresentation/builtInToolPresentation.ts
