+++ src/frontend/components/messagePresentation/utils/builtInToolPresentation/builtInToolPresentation.types.ts
