--- src/frontend/features/chat/input/components/ChatInputEffortBackdrop.tsx
@@ -1,2 +1,0 @@
-export { ChatInputEffortBackdrop } from './ChatInputEffortBackdrop.android';
-export type { ChatInputEffortBackdropProps } from './chatInputEffortBackdrop.types';
