+++ src/frontend/features/chat/input/components/ChatInputEffortBackdrop/ChatInputEffortBackdrop.tsx
@@ -1,0 +1,2 @@
+export { ChatInputEffortBackdrop } from './ChatInputEffortBackdrop.android';
+export type { ChatInputEffortBackdropProps } from './ChatInputEffortBackdrop.types';
