+++ src/frontend/features/chat/input/components/ChatInputEffortBackdrop/ChatInputEffortBackdrop.types.ts
